Add files via upload
##How to get started

To make a move, the player will use the mouse single click to mark a space. In this version, there is no provision to undo the move. As soon as the move is done, the game proceeds to the next player’s chance.
At each move, The game will show whose move (Player A or Player B) it is. When the game ends, it displays one of the following three outcomes:
Winner: Player A
Winner: Player B
Draw